// frontend/src/client/models/Experience.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Experience = {
    company?: (string | null);
    logo?: (string | null);
    position?: (string | null);
    description?: (string | null);
    start_date?: (string | null);
    end_date?: (string | null);
};